partial search and bigdecimal conversion

+++ order-service/src/main/java/com/ecommerce/order/dto/OrderItemResponse.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public record OrderItemResponse(
         Long productId,
         Integer quantity,
-        BigDecimal priceAtPurchase,
-        BigDecimal itemSubtotal
+        Long priceAtPurchase,
+        Long itemSubtotal
 ) {}
