+++ payment-service/src/main/java/com/ecommerce/payment/dto/PaymentRequest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record PaymentRequest(
         Long orderId,
-        BigDecimal amount,
+        Long amount,
         String customerEmail,
         String productName
 ) {
